Fixed up NPC collision box and added dialogue after starting quest plus added objectives animation.

+++ Assets/Scripts/PlayerData.cs
@@ -9,5 +9,5 @@ public class PlayerData : MonoBehaviour
     // for the RPG, we'll need a variable to store which quest the player is on atm.
     // quest number 0 is the default and it means that they're not on an actual quest.
     public int questNumber;
-    public int dialogueNumber;
+    public float dialogueNumber;
 }
